requirements.txt
django
numpy
pandas
openai
tqdm
names
djangorestframework
bcrypt
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
django-cors-headers
gunicorn
psycopg
django-crypto-fields>=0.4.2
spacy
presidio_analyzer
presidio_anonymizer
django-phonenumber-field[phonenumbers]
edc_constants
django-extensions
Werkzeug
pyOpenSSL
# python -m spacy download en_core_web_lg